Cleaned up atom.h
